+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -225,6 +225,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsMock()
      */