+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -185,6 +185,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertIsArrayNotEmpty()
      */