+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -172,6 +172,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertImageSize()
      */