+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -149,6 +149,7 @@ class TestTraitTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses \Tools\TestSuite\TestTrait::assertFileExtension()
      */