+++ tests/TestCase/GlobalFunctionsTest.php
@@ -137,6 +137,7 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses uncamelcase()
      */