+++ tests/TestCase/GlobalFunctionsTest.php
@@ -78,6 +78,7 @@ class GlobalFunctionsTest extends TestCase
     }
 
     /**
+     * @group legacy
      * @test
      * @uses is_html()
      */