+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.9 branch
 ### 1.9.4
 * `TestTrait::assertFileMime()` is deprecated and will be removed in a future release;
+* `TestTrait::assertImageSize()` is deprecated and will be removed in a future release;
 * `TestTrait::assertIsArrayNotEmpty()` is deprecated and will be removed in a future release;
 * `TestTrait::assertIsMock()` is deprecated and will be removed in a future release;
 * `TestTrait::assertObjectPropertiesEqual()` is deprecated and will be removed in a future release;